+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Publish -SNAPSHOT releases to Maven Central. (CIS-3368)
+
 ## [0.2.0] - 2025-09-30
 
 ### Added