+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.0.0] - 2025-10-21
+
 ### Added
 
 - Added the method 'prepare' to NotificationViewer to allow applications to optimize performance. Default does nothing.
@@ -68,7 +70,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Use Spring Boot Autoconfiguration (CIS-3347)
 
-[unreleased]: https://github.com/OHSU-OCTRI/notification-lib/compare/v0.3.0...HEAD
+[unreleased]: https://github.com/OHSU-OCTRI/notification-lib/compare/v1.0.0...HEAD
+[1.0.0]: https://github.com/OHSU-OCTRI/notification-lib/compare/v0.3.0...v1.0.0
 [0.3.0]: https://github.com/OHSU-OCTRI/notification-lib/compare/v0.2.0...v0.3.0
 [0.2.0]: https://github.com/OHSU-OCTRI/notification-lib/compare/v0.1.1...v0.2.0
 [0.1.1]: https://source.ohsu.edu/OCTRI-Apps/compass/compare/v0.1.0...v0.1.1